refactoring of project files with better naming and organisation
