+++ Synto/ml/training/__init__.py
@@ -1,7 +1,7 @@
 from .loading import *
-from .policy_training import *
+from .supervised import *
 from .preprocessing import ValueNetworkDataset, mol_to_pyg, MENDEL_INFO, compose_retrons
-from .policy_training import create_policy_training_set, run_policy_training
+from .supervised import create_policy_training_set, run_policy_training
 
 __all__ = [
     "ValueNetworkDataset",
